c.snippets: main() const argc (argv untouched)
Due to common pratice of representing argv as an array, we can't make
argv const in any way without changing its representation. Do note that
arrays as function parameters are actually pointers, so we could
technically declare argv as char** const to achieve constness of argv as
described, but I'd rather leave it denoted as an array.