vim: nowrap; fix how cursorline is applied
Using BufWinEnter and setlocal can make cursorline apply to Telescope's
preview window, for example. This isn't ideal, as it'll always be seen
on the first line.

Reason for this is explained in :h local-options.
:setlocal on window-local options applies it to the window, but only for
the buffer displayed in that window. Set the global value with :set,
which applies it to the window, but for all buffers that may display
within it.

Also adjust some comments.