nvim/treesitter: disable highlights in some cases, update gps icons
vim's treesitter highlights are a little wonky atm, use the built-in syntax.
Also disable highlights for largish files (>3000 lines).

I'm not a fan of using special font icons as a default; I'd rather replace gps
in the future if I need to keep up with additions like this (also, this doesn't
even consider the language-special icons like with json).