vim/nvim: remove vsnip, friendly-snippets
I rarely use snippets, even when I try to force myself to!
Even with the completion I added, I still find it hard to figure out what
snippets are available, so friendly-snippets doesn't help much.

The only reason keeping snippet support *might* be useful is for LSP completions
(some completions use snippets for filling in and jumping around function
params, etc.).
As I only use LSP in Nvim, it might be worth using LuaSnip in that case.

For now, it might be nice to show signature help when LSP completing...