nvim: add neoformat key mapping
Originally removed the old key mapping after transitioning away from function
key mappings, but typing `:Neo<CR>` all the time is more annoying, and causes
hit ENTER prompts with &ch=0.

Remains to be seen how effective this mapping is, as I've been away for a while;
will be back soon... whenever that is.