nvim: leave document_colors enabled by default
Nvim now enables documentColors *for all clients attached to the buffer* by
default, but does it *every time a new client attaches*; this makes managing it
via a toggle annoying, as a new client attaching to the buffer can make the
current toggled state ineffective unless we keep track of it ourselves.

Instead of doing that, just embrace the default for now. It's likely better to
just disable documentColors for particular servers that overuse them?