nvim: give each client a progress id, don't schedule
This means a progress update is given 1:1 for each progress update from a
server. It's moreso appropriate for the progress implementation to handle them
anyway.

Currently ui2 doesn't handle progress messages too well (they don't necessarily
replace the last message in the command-line); the legacy message grid handles
it better, but will spam press-ENTER prompts if the screen width is too small.
Plus there's no way to disable showing it in the command-line. So this is still
waiting on upstream improvements.